+++ NEWS.md
@@ -1,3 +1,8 @@
+# leunbachR 0.0.2 (2026-01-20)
+
+- Theta scores added to output.
+- Vignette describes how to export bootstrap equating table to CSV file.
+
 # leunbachR 0.0.1 (2026-01-15)
 
 - Initial release